+++ packages/dropdown/src/DropdownContent.tsx
@@ -10,7 +10,13 @@ export function DropdownContent({ children, ...props }: DropdownContentProps) {
   return (
     <DropdownMenu.Portal>
       <DropdownMenu.Content sideOffset={4} asChild {...props}>
-        <Card direction="column" padding="8px 0" minWidth={192} maxWidth={320}>
+        <Card
+          direction="column"
+          padding="8px 0"
+          minWidth={192}
+          maxWidth={320}
+          css={{ zIndex: 5 }}
+        >
           {children}
         </Card>
       </DropdownMenu.Content>