+++ packages/form-components/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @siakit/form-components
 
+## 0.0.3
+
+### Patch Changes
+
+- add pagination, table and modify other components
+
 ## 0.0.2
 
 ### Patch Changes